requirements.txt
# Core dependencies
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
upstash-vector>=0.6.0

# Development and testing
httpx>=0.25.0